Rewrite disabled track handling to always load MIDI
Since updates are infrequent, and creating tracks is intended to be
cheap, we can afford to create and then immediately destroy tracks here.

This approach ensures that we always calculate the full song duration,
even when all tracks are disabled. However, I didn't want to keep
unnecessary MIDI players running after the full transport has been
loaded, so we now reference count the MIDI players, removing unnecessary
ones after the disabled tracks are cleaned up.

If a song is playing with no enabled tracks, there will be no
MIDIPlayers running, which means that we can no longer rely on the
MIDIPlayer endCallback to trigger song completion. This changeset also
switches to setting a timeout after the remaining time on the song
finishes.